backend/requirements.txt
Flask==2.2.5
requests==2.28.1
python-dotenv==0.20.0
Flask-SQLAlchemy==2.5.1
httpx==0.23.1